+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 ### Added
 ### Fixed
 
+## 0.4.4 - 2025-07-01
+
+### Fixed
+- Make GPG config be consistently ordered in lockfile.
+- Remove debug print statement.
+- Handle case where RPMs have a different checksum algorithm than expected when they're being downloaded, while still verifying the checksum.
+
 ## 0.4.3 - 2025-05-28
 
 ### Added
@@ -32,7 +39,7 @@
 
 ### Fixed
 - Set size in hardlink headers correctly.
-  - Fixes integrity failures during this `docker push` 
+  - Fixes integrity failures during this `docker push`
 
 
 ## 0.3.1 - 2024-07-24